FIX The extrafield factor_depot must be editable when invoice is validated
The extrafield is create without attribute "edit whatever is status".
So when invoice is validated, it is no more possible to set the field factor_depot forcing user to unvalidate the invoice and then validate it. This is bad because we update this field in most cases after having validated the invoice.